refactor: Session, SessionManager 구조 변경
- isNew 메서드로 인해 Session SessionManager 양방향 의존 해결이 필요했음
- Session이 항상 SessionManager에서 만들어지도록 보장하면 로그인 여부 확인 시
isNew 메서드가 필요없는 상황
- 따라서 SessionManager에서 항상 Session을 만들도록 함
- 외부의 Session 생성 방지를 위해 내부 클래스로 변경
